Waiter Side

+++ app/models/models.py
@@ -57,7 +57,7 @@ class Order(db.Model):
     table_id = db.Column(db.Integer, db.ForeignKey("tables.id"), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     status = db.Column(db.String(20), default="pending")
-    total_amount = db.Column(db.Numeric(10, 2))
+    total_amount = db.Column(db.Numeric(15, 2))
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
     updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
 
@@ -70,7 +70,7 @@ class OrderItem(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     order_id = db.Column(db.Integer, db.ForeignKey("orders.id"), nullable=False)
     menu_item_id = db.Column(db.Integer, db.ForeignKey("menu_items.id"), nullable=False)
-    quantity = db.Column(db.Integer, default=1)
+    quantity = db.Column(db.Numeric(5, 2), default=1.00)
     price = db.Column(db.Numeric(10, 2), nullable=False)
     vip_price = db.Column(db.Numeric(10, 2))  # VIP price support
     notes = db.Column(db.Text)
